requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3     
billiard==4.2.0
celery==5.3.4
certifi==2023.11.17      
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0  
click-plugins==1.1.1     
click-repl==0.3.0        
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.6
django-celery-beat==2.5.0
django-redis==5.4.0
django-templated-mail==1.1.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
flower==2.0.1
humanize==4.8.0
idna==3.6
kombu==5.3.3
oauthlib==3.2.2
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.9